0919 cho - ProductCategory에서 커스텀 훅, 모달 분리
